+++ CHANGES.md
@@ -4,6 +4,9 @@ Notable changes between releases.
 
 ## Latest
 
+* Add `expression` field to `cue_config` ([#3](https://github.com/poseidon/terraform-provider-cue/pull/3))
+  * Evaluate an expression instead of an entire config
+
 ## v0.1.0
 
 * Add `cue_config` data source to evaluate CUE contents